FileUpload/requirements.txt
Django==2.2.13
djangorestframework==3.11.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
